Base project setup; base Window, base menubar, and base tabs
Pirex.java contains main() which calls for the construction of the
window and the contents inside it.

PirexWindow.java creates frame, calls constructor MainInterface(), sets
up a basic menu bar with an ActionListener.

MainInterface.java essentially all of the user interface that is no the
menu bar or the window. Creates a JTabbedPane with 3 different tabs.

Pirex.png Not currently used but will be the logo we use for the title
section of the window.