Version: 1.2.1 Fixed application Icon to work on any machine
files updated: PirexWindow.java; Pirex.java

fixed Icon to direct to a path that does not rely on any folders
besides the one where the application is stored.

Added notes to grader section.

added functionality for Help > Index to load a file that will offer
information on how to operate the application